frontend/requirements.txt
# Core dependencies
google-generativeai>=0.3.0
requests>=2.31.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
scipy>=1.11.0
pyttsx3>=2.90
python-dotenv>=1.0.0

# MoviePy and video processing (install these in order)
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
decorator>=4.4.2
proglog>=0.1.10
tqdm>=4.66.0
moviepy==1.0.3
